+++ HISTORY.rst
@@ -9,6 +9,8 @@ As of 1.0.0, `semantic versioning <http://semver.org/>`__ is used.
 +++++++++
 released 2013-XX-XX
 
+- add Mobileclient.get_album_info
+
 
 2.0.0
 +++++